Bug fix with imgFrame copy when convolving vmap with PSF.
Fixed bug where first the PSF kernel image was not extended far enough which gave zeros in the outer parts of the PSF-convolved velocity map. Fixed that part by adding imgFrame when drawing the object to specify the size in pixels. But then this caused a problem where the same image frame was filled multiple times because of how python assigns pointers instead of copying objects by default to save memory.

Also added a bunch of comments, changed gal_beta from radians to degrees, handled case of sigma=0 when weighting in vmapFit.




git-svn-id: svn+ssh://riemann.lbl.gov/home/mgeorge/svn/speclens@7 71aa9303-448b-4a3d-81da-bc994458fe86