Added g1,g2 as parameters to vmapModel.
Now pars is a 5-element array used to build a vmap. It does this by taking a set of observed angles (xvals), de-shearing them to put them into the source plane, and computing the vmap as usual.

Updated interpretPriors to add g1 and g2 to guess and guessScale, and fixed how it handles fixed parameters (this hasn't been thoroughly tested yet).



git-svn-id: svn+ssh://riemann.lbl.gov/home/mgeorge/svn/speclens@19 71aa9303-448b-4a3d-81da-bc994458fe86