Added imaging observables to fit.
vmapModel now returns model fiber velocities as well as sheared ellipse b/a and PA. lnProbVMapModel now considers both velocity and imaging observables, and compares that data to the model to compute a likelihood. vmapFit now provides those observables to lnProbVMapModel.

There is an option to set either type of observable to None to exclude it in the fitting.

Also added saving to contourPlotAll, and change some variable names for clarity.


git-svn-id: svn+ssh://riemann.lbl.gov/home/mgeorge/svn/speclens@23 71aa9303-448b-4a3d-81da-bc994458fe86