Added 2 significant features - generateEnsemble and getFiberFluxes
generateEnsemble creates a list of pars given a set of priors. This can be used for making an ensemble of galaxy properties with a set of shears applied, to test how a distribution of shapes and shears can be fit.

getFiberFluxes is intended to replace getFiberFlux and the integrand functions that it calls, but is still under testing. The idea is to just do k-space convolutions on the galaxy image and velocity map (for the PSF and for the fiber area), then sample this convolution at the center of the fiber in real-space to get the fiber flux. This should perform much faster than the real-space integral over the fiber area currently used, and seems to give fairly consistent results. Also avoids issues with real-space convolution (for the PSF) in galsim, which is currently being debugged.




git-svn-id: svn+ssh://riemann.lbl.gov/home/mgeorge/svn/speclens@32 71aa9303-448b-4a3d-81da-bc994458fe86