moved main from ensemble.py and ensemblePlots from makeplots.py to runEnsemble.py
Fixes #5 and creates a cleaner example of the functions in ensemble.py
ensemble.py had a mix of general use modules and user-specific ones
like create_qsub_galArr which was for submitting galaxy jobs in
parallel to a queue scheduler. Some of that code remains in
scratch/tf_ensemble.py for now, but it has been removed from
ensemble.py.

Previously the final example in makeplots.py and the main program in
ensemble.py used different approaches to generate and fit 100
galaxies. The new module in examples/ uses the functionality built in
ensemble.py (without the queue job submission) to run 100 galaxies
serially. This is currently quite slow but illustrates the approach
for running an ensemble of galaxies.

Also cleaned up noise handling in makeObs allowing None args for sigma
and ellErr, though those args wouldn't be useful in fitting.