started base.py to define Galaxy and Observation classes
To collect the various parameters used to define a galaxy and to
better separate the functions that generate galaxies from those that
produce observables, define Galaxy and Observation classes. Future
additions may define a Fit class that can centralize model definitions
and priors.

Currently, Galaxy class just collects the variables used to define the
intrinsic properties of a galaxy. Observation class takes a Galaxy
object and all the additional parameters that define an observer
(redshift, position angle, etc) and has methods to produce
observables. These are initially just wrappers to functions in sim.py,
and it may be helpful to retain the existing approach with arrays of
variables to define galaxies and image arrays, but as the list of
variables gets longer (with disk thickness, see e.g. Issue #8), the
object approach should be tidier.