add convertInclination function
should supercede getInclination since it allows for thick disks,
taking any two args of galBA, galCA, and inc and returns the 3rd. So
getInclination should be equivalent to convertInclination when
inc=None, galCA=0, and galBA=gal_q. Still TBD is to clarify whether
galBA is the intrinsic or observed axis ratio, and how to switch
between the two.