Add Model object in base.py
Starting from a suggestion in #8, build a model object as the
clearinghouse for different model definitions and parameter
arrangments. Vision is that all parameters needed to construct a
galaxy and observables will be collected in base (currently there is
some redundancy defining these in both the Galaxy/Observation objects
and in the Model object - these could be unified eventually). Model
will have some methods borrowed from interpretPriors in fit.py to
handle fixing parameters and setting priors, and will eventually
remove the need for an ordered list of pars (except the one sent to
emcee) in all sim.py methods.