add model 'B' to base.py Model class, adding thickness to pars list
This mostly columinates work on #8 to build a more flexible Model
class which should enable easier switching between model
parametrizations. In this case, model B extends model A by adding
diskCA to the array of fit parameters. In both cases the full model
object is passed around between the functions in sim.py that build the
galaxy and velocity map, but the class method updatePars handles the
appropriate list of parameters to update with each set of par guess
from an MCMC chain. Future models can be added (with more descriptive
names) by simply adding blocks to Model.init and Model.updatePars.