replace diskBA with cosi in Model A and B parameter lists
With the introduction of disk thickness (diskCA != 0), it becomes
unphysical to have the projected image axis ratio (diskBA) take
certain values. For example, if diskCA=0.2, the minimum value that
diskBA can take is 0.2 (the edge on case) - the projected disk can be
no thinner. This means that the prior on diskBA is dependent on
diskCA, and if values are guessed randomly then unphysical results may
occur. If instead the disk inclination is used (or rather cos(i),
which equals diskBA for thin disks), then it can be assigned randomly
and indepently from diskCA. Randomly oriented disks will have a
distribution of cos(i) that is uniform from 0 to 1, so a simple prior
can be kept.

This commit replaces diskBA with cosi in the pars list by changing
only the model definition and updatePars method in base.py. While cosi
will be the value in the pars list used by emcee, updatePars
immediately assigns the relevant value to diskBA given cosi and
diskCA, producing self-consistent, physical model galaxies. Useful for
issue #8.