replace galsim construction with makeImageBessel in makeGalVMap2
makeGalVMap2 (used to construct galaxy images and velocity maps when
convOpt=pixel, which is now the default option) now generates an image
array based on model pars using makeImageBessel which follows Spergel
2010 and approximates empirical surface brightness profiles with
Bessel functions. This allows easy generation of thin and thick disks
with the same inclination and PA.

The flux-weighting for the velocity map now uses a thin disk (assuming
emission line flux comes from a thin disk) which otherwise has the
same pars as the model object which describes the (thick and possibly
bulgy) stellar image. makeGalVMap2 now also returns this thin disk
image array, so that the flux in a given vsample can be computed to
properly normalize the velocity after beam convolution.

This change brings updates to makeBesselImage, including a bug fix in
the port from the IDL version (there's only one sqrt in L504) and
rotates phi_r by 90 deg so that the PA of the image agrees with the
convention used to generate the vmap (i.e. PA=0 corresponds to
x,y=1,0) and positive is CCW).

Update mostly wraps up #8, next step is merging with main branch.