enable parallelization of emcee
Fixes #11 (without directly addressing #15) To allow multiprocessing
with emcee, just need to give EnsembleSampler a number of threads to
use, which can now be passed down from ensemble.runGal to fit.fitObs
to fit.vmapFit.

The tricky part was that the likelihood function and its argument list
need to be pickleable, and the model object after inclusion of
priorFuncs was not. The workaround is to modify interpretPriors to
return priorFuncs rather than write it as an attribute to model. And
the likelihood function (lnProbVMapModel) now calls interpretPriors
again (in each iteration) to get priorFuncs. priorFuncs and priors
still ought to be revamped (see #15), but this step works for
parallelization.