remove disk rescaling in shearEllipse, add defineEllipse
shearEllipse was changing disk_r by a factor of (1+gamma), which isn't
a proper treatment of magnification. As far as I can tell, this only
impacted imaging visualization because disk_r wasn't used in ellModel
for fitting (and it was hidden from the visualization tests by a fudge
that made fig1b and fig1c in makeplots.py identical)

defineEllipse is a small helper function to produce an ellipse array
from model attributes.