testing alternative shearEllipse
When running a long chain, sometimes a set of pars is guessed which
produces unphysical results for sheared ellipses - the computed
ellipticity or axis ratio is out of the range 0-1. This is presumably
due to weak shear approximations breaking down when large shears are
guessed. But the eqs from Supri & Harari 1999 used in shearEllipse
still need testing, and do not appear to always be consistent with
those in Huff++.

Commit starts with new method convertShear (#7) to try switching
between shear coordinate systems, and adds shearEllipse2 which is
meant to codify eqs in Huff++ for comparison with shearEllipse.

Still needs testing and clarification since it occasionally produces
negative axis ratios after shear.