start TestObservable class, to replace Observation
TestObservable, to eventually be called Observable, holds data that
could be obtained from a telescope.

Context: the likelihood code computes (model - data) / error.  To
enable a flexible range of data types (imaging, velocities, full
datacubes), the observable class will hold data and a description of
what form the model should be projected for comparison to the data. So
ultimately the Model class will have methods to generate data in these
different formats (which can be saved as simulated Observable objects)
and the likelihood code can compare those objects to the actual data
(saved in the same type of Observable objeect).

A step towards #17