use Bartelmann & Schneider for shearEllipse
Replace old shearEllipse with updated shearEllipse3, and remove
shearEllipse2 which is more of an approximation from 1311.1489 which
has a sign error or different convention on the axis ratio
transformation.

Also start the test suite, replacing test_makegal with test_sim which
has some initial tests of the new shearEllipse and ellModel methods.