develop new model, observable, detector classes in scratch
Working toward object-based definitions of intrinsic galaxy properties
(TestGalaxy), observable data (TestObservable) which has detector info
(TestDetector), and a model class that holds an array of parameters
along with the full set of intrinsic galaxy properties and simulated
observables (TestModel, which holds a TestGalaxy and TestObservable).

Ultimately, this will make it easier to compare model to data for
likelihood calculation, with proper self-description in data space for
observed PA wrap issue (#16) and flexible handling of different
observable types (#17).