use @property decorators to generate diskBA and rotCurvePars on the fly
In Model class, diskBA and rotCurvePars are useful to store as
attributes, but they depend on the values of other attributes. Paste
versions tried to update diskBA whenever diskCA or cosi are changed,
but really we want diskBA to be generated on the fly with whatever the
current values of its parent attributes are.

The property decorator enables this. See
http://adam.gomaa.us/blog/2008/aug/11/the-python-property-builtin/ for
use cases. This shouldn't change the syntax for any other function
trying to access the model.diskBA attribute (e.g. the getter is just
model.diskBA), but it will affect the behavior of any code that tries
to set diskBA. We don't want to set the intrinsic diskBA directly
though.