save model and observation attributes to text file
add writeObj function to io.py which prints the string representation
of an object (like a model or observation) to a text file, one line
per dict item.

add call to writeObj in ensemble.runGal. Also clean up the calls to
write chain files in that function with a for loop.